Modify: mock data 활용, Refactoring 진행중
