+++ src/pages/seonghunEom/Login/Login.js
@@ -7,39 +7,51 @@ const Login = () => {
   const postUserData = () => {
     navigate('/main-seonghun');
   };
-  const [id, setId] = useState('');
-  const checkId = e => setId(e.target.value);
-  const [pw, setPw] = useState('');
-  const checkPw = e => setPw(e.target.value);
+
+  const [inputValue, setInputvalue] = useState({
+    userId: '',
+    userPassword: '',
+  });
+  const { userId, userPassword } = inputValue;
+
+  const handleInput = e => {
+    const { name, value } = e.target;
+    setInputvalue({ ...inputValue, [name]: value });
+  };
+
   const [disable, setDisable] = useState(true);
   const handleBtn = () => {
-    id.includes('@') && pw.length >= 5 ? setDisable(false) : setDisable(true);
+    userId.includes('@') &&
+    userId.includes('.') &&
+    userId.length >= 5 &&
+    userPassword.length >= 5
+      ? setDisable(false)
+      : setDisable(true);
   };
+
   return (
     <div className="login">
       <div className="loginCard">
         <h2>WestagramLogin</h2>
-        <form className="form">
+        <form className="form" onKeyUp={handleBtn} onChange={handleInput}>
           <label className="idLabel">
             <input
+              name="userId"
               type="text"
               placeholder="전화번호, 사용자 이름 또는 이메일"
               className="id"
-              onChange={checkId}
-              onKeyUp={handleBtn}
             />
           </label>
           <label className="idLabel">
             <input
+              name="userPassword"
               type="password"
               placeholder="Password"
               className="pw"
-              onChange={checkPw}
-              onKeyUp={handleBtn}
             />
           </label>
           <button
-            type="button"
+            type="submit"
             className="submitBtn"
             disabled={disable}
             onClick={postUserData}
