+++ src/pages/seonghunEom/Main/Main.scss
@@ -1,303 +1,28 @@
 @charset "utf-8";
 
-@mixin flexCenter {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.main {
+.main-wrap {
   background-color: #fafafa;
   margin: 0 auto;
 
-  .main-navigation {
-    position: fixed;
-    width: 100%;
-    border-bottom: 1px solid #e6e6e6;
-    background-color: #fff;
-    font-family: ‘Noto Sans KR’, sans-serif, ’dotum’, ’돋움’,
-      ’Apple SD Gothic Neo’;
-
-    .toLogin {
-      @include flexCenter;
-      text-decoration: none;
-      color: inherit;
-      font-family: ‘Noto Sans KR’, sans-serif, ’dotum’, ’돋움’,
-        ’Apple SD Gothic Neo’;
-      background-color: #4a9bd4;
-      padding: 8px 0;
-      font-weight: bold;
-      color: #ffffff;
-    }
-  }
-
-  .nav-box {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 1000px;
-    height: 50px;
-    margin: 0 auto;
-
-    .logo {
-      display: flex;
-
-      i {
-        margin-right: 10px;
-        font-size: 25px;
-      }
-
-      h1 {
-        font-size: 25px;
-      }
-    }
-  }
-
-  .search-box form label input {
-    width: 200px;
-    text-align: center;
-    height: 30px;
-    outline: none;
-    border: 1px solid #dbdbdb;
-    border-radius: 5px;
-    background-color: #fafafa;
-    // color: #d5d5d5;
-
-    input::placeholder {
-      color: green;
-    }
-  }
-
-  .nav-icons a {
-    margin-right: 15px;
-  }
-
-  .nav-icons a:last-child {
-    margin-right: 0px;
-  }
-
-  .nav-icons a img {
-    width: 25px;
-  }
-
-  main {
+  .main {
     width: 950px;
     margin: 0 auto;
     display: flex;
     justify-content: space-between;
     padding-top: 120px;
+    padding-bottom: 120px;
 
     .feeds {
       width: 550px;
       height: auto;
-      background-color: #fff;
-      border: 1.5px solid #e6e6e6;
       display: flex;
       flex-direction: column;
-      justify-content: center;
       align-items: center;
-      margin-bottom: 100px;
-
-      article {
-        width: 550px;
-        height: auto;
-
-        .profiles {
-          display: flex;
-          align-items: center;
-          padding: 15px 18px;
-
-          .profiles-img {
-            margin-right: 10px;
-            font-weight: bold;
-            width: 40px;
-            height: 40px;
-            overflow: hidden;
-            border-radius: 50%;
-
-            a img {
-              width: 100%;
-              height: 100%;
-              object-fit: cover;
-            }
-          }
-
-          .id-text {
-            font-weight: bold;
-            color: #414040;
-          }
-
-          .article-header-menu {
-            margin-left: auto;
-
-            i {
-              color: #383838;
-            }
-          }
-        }
-      }
-
-      .contents {
-        width: 100%;
-        margin: 0 auto;
-        height: 600px;
-        overflow: hidden;
-        border: 1px solid #e6e6e6;
-        outline: 1px solid #e6e6e6;
-
-        img {
-          width: 100%;
-          height: 100%;
-          object-fit: cover;
-        }
-      }
-
-      .article-bottom {
-        width: 100%;
-
-        .article-links {
-          display: flex;
-          justify-content: space-between;
-          width: 100%;
-          padding: 10px;
-
-          i {
-            font-size: 23px;
-            font-weight: 100;
-            width: 23px;
-            height: 23px;
-            margin-right: 10px;
-            color: #414141;
-          }
-
-          i:last-child {
-            margin-right: 0;
-          }
-        }
-
-        .article-links-right i {
-          font-size: 23px;
-          font-weight: normal;
-          width: 23px;
-          height: 23px;
-        }
-
-        .article-bottom-like {
-          width: 100%;
-          padding: 10px;
-          font-size: 15px;
-
-          .like-profile {
-            display: flex;
-            align-items: center;
-
-            .like-profile-myImg {
-              display: inline-block;
-              width: 20px;
-              height: 20px;
-              overflow: hidden;
-              border-radius: 50%;
-              margin-right: 8px;
-
-              img {
-                width: 100%;
-                height: 100%;
-                object-fit: cover;
-              }
-            }
-          }
-
-          .article-bold {
-            font-weight: bold;
-          }
-
-          .article-bold-child {
-            font-weight: bold;
-          }
-        }
-
-        .article-bottom-texts {
-          padding: 10px;
-          width: 100%;
-          font-size: 15px;
-
-          .article-bold {
-            font-weight: bold;
-          }
-        }
-
-        .article-bottom-texts-new {
-          display: flex;
-          justify-content: space-between;
-
-          i {
-            margin-left: 6px;
-          }
-        }
-
-        .article-bottom-texts > div {
-          margin-bottom: 5px;
-        }
-
-        .upload-time span {
-          color: #9a9a9a;
-        }
-
-        .article-bottom-comments {
-          width: 100%;
-          padding: 10px;
-          border: 1px solid #dbdbdb;
-
-          form {
-            display: flex;
-            align-items: center;
-
-            .emotion {
-              width: 5%;
-              height: 20px;
-
-              i {
-                font-size: 20px;
-              }
-            }
-          }
-
-          textarea {
-            width: 85%;
-            height: 30px;
-            outline: none;
-            border: none;
-            resize: none;
-            line-height: 25px;
-            text-indent: 5px;
-          }
-
-          button.comments-submit-false {
-            width: 10%;
-            height: 30px;
-            outline: none;
-            border: none;
-            background-color: #f0f6ff;
-            border-radius: 10px;
-            font-weight: bold;
-          }
-
-          button.comments-submit-true {
-            width: 10%;
-            height: 30px;
-            outline: none;
-            border: none;
-            background-color: #f0f6ff;
-            color: #3d87ea;
-            border-radius: 10px;
-            font-weight: bold;
-          }
-        }
-      }
     }
 
-    /* aside */
     aside {
+      position: sticky;
+      top: 70px;
       width: 350px;
       height: 700px;
       background-color: #fafafa;
